requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary  # For PostgreSQL
dj-database-url
whitenoise
drf_yasg